refactored dialogs to separate folder
